Core, Vulkan, Platform: minor warning fixups

+++ Hermes/Source/Core/Misc/StringUtils.cpp
@@ -13,7 +13,7 @@ namespace Hermes
 	{
 		String Result(In.size(), 0);
 #ifdef HERMES_PLATFORM_WINDOWS
-		MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, In.c_str(), -1, Result.data(), Result.size());
+		MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, In.c_str(), -1, Result.data(), (int)Result.size());
 #else
 		// TODO : implement properly
 		// std::wstring seems not to be completely cross-platform, we should deal with it when we actually move to Linux systems
@@ -25,7 +25,7 @@ namespace Hermes
 	{
 		ANSIString Result(In.size(), 0);
 #ifdef HERMES_PLATFORM_WINDOWS
-		WideCharToMultiByte(CP_ACP, 0, In.c_str(), -1, Result.data(), Result.size(), NULL, NULL);
+		WideCharToMultiByte(CP_ACP, 0, In.c_str(), -1, Result.data(), (int)Result.size(), NULL, NULL);
 #else
 		// TODO : unimplemented
 #endif
